some more checks for int overflows
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@1857609 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/ss/formula/EvaluationConditionalFormatRule.java
@@ -455,7 +455,7 @@ public class EvaluationConditionalFormatRule implements Comparable<EvaluationCon
                         Collections.sort(allValues);
                     }
                     
-                    int limit = (int) conf.getRank();
+                    int limit = Math.toIntExact(conf.getRank());
                     if (conf.getPercent()) {
                         limit = allValues.size() * limit / 100;
                     }
